move spawn tests from test_c10d.py to test_c10d_spawn.py
